Began work on the signup activity and java class
