+++ app/src/main/java/org/wit/mytweet/activities/SignUp.java
@@ -1,0 +1,16 @@
+package org.wit.mytweet.activities;
+
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import org.wit.mytweet.R;
+
+public class SignUp extends AppCompatActivity{
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_signup);
+    }
+}
