+++ app/src/main/res/layout/activity_launch.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.wit.mytweet.LaunchActivity"
+    tools:context="org.wit.mytweet.activities.LaunchActivity"
     android:background="@drawable/gradient">
 
     <TextView
@@ -22,15 +22,17 @@
 
     <TextView
         android:id="@+id/launchMessage"
+        android:gravity="center"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/launchMessage"
-        android:layout_marginTop="32dp"
-        app:layout_constraintTop_toBottomOf="@+id/welcomeMessage"
         android:layout_marginLeft="16dp"
-        app:layout_constraintLeft_toLeftOf="parent"
         android:layout_marginRight="16dp"
-        app:layout_constraintRight_toRightOf="parent" />
+        android:layout_marginTop="32dp"
+        android:text="@string/launchMessage"
+        android:textAppearance="@android:style/TextAppearance.Material.Medium"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/welcomeMessage" />
 
     <Button
         android:id="@+id/signupButton"
@@ -49,9 +51,7 @@
         android:text="@string/loginButton"
         android:layout_marginTop="128dp"
         app:layout_constraintTop_toBottomOf="@+id/launchMessage"
-        android:layout_marginRight="0dp"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintLeft_toRightOf="@+id/signupButton"
-        android:layout_marginLeft="8dp" />
+        android:layout_marginRight="64dp"
+        app:layout_constraintRight_toRightOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
